chore/#225 : 리스트 목록 뷰 한 파일로 일단 합침
