+++ components/folder/FolderLanding.tsx
@@ -212,7 +212,7 @@ function FolderLanding() {
   };
 
   const handleFolderClick = (id: string, categoryName: string) => {
-    router.push(`/packing-list/${categoryName}?id=${id}`);
+    router.push(`/packing-list?type=${categoryName}&id=${id}`);
   };
 
   const getSwiperIndex = (index: number) => {
