+++ pages/packing-list/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import AlonePackingListLanding from '../../components/packingList/alone/AlonePackingListLanding';
+import PackingListLanding from '../../components/packingList/PackingListLanding';
 import { AsyncBoundary } from '../../utils/AsyncBoundary';
 
 function index() {
   return (
     <AsyncBoundary>
-      <AlonePackingListLanding />
+      <PackingListLanding />
     </AsyncBoundary>
   );
 }
